The march forward

+++ src/xkb/keysym.rs
@@ -21787,7 +21787,7 @@ pub mod keysym_names_h {
             explicit_count: 0 as u8,
         },
     ];
-    pub static mut explicit_deprecated_aliases: [u32; 1] = [24103 as u32];
+    pub static mut explicit_deprecated_aliases: [u32; 1] = [24103];
 }
 pub use crate::xkb::utf8_decoding::{utf8_next_code_point, INVALID_UTF8_CODE_POINT};
